refactor(naming): share k8s_safe_name via nemo_platform_plugin (AIRCORE-860)
Extract DNS-safe naming into nemo_platform_plugin.k8s_naming so deployments
and models share one contract. Use DNS_LABEL_MAX_LENGTH constant, delegate
puller naming to get_docker_puller_container_name, and document edge-case
fallbacks per review feedback.
