+++ plugins/nemo-deployments/README.md
@@ -52,5 +52,7 @@ as `network`.
 Docker container and volume names use a readable prefix plus a deterministic
 8-character hash suffix. The hash is computed from ``{workspace}/{name}``, not
 from the hyphen-joined string, so pairs like ``foo``/``bar-baz`` and
-``foo-bar``/``baz`` cannot collide. Orphan cleanup matches identity labels, not
-names alone; existing containers keep their old names after upgrade.
+``foo-bar``/``baz`` cannot collide. Naming logic is shared via
+``nemo_platform_plugin.k8s_naming`` (same module used by the models service).
+Orphan cleanup matches identity labels, not names alone; existing containers
+keep their old names after upgrade.
