+++ plugins/nemo-deployments/src/nemo_deployments_plugin/backends/docker/labels.py
@@ -9,17 +9,15 @@ Resource names are derived from a human-readable prefix plus a deterministic
 pairs like ``("foo", "bar-baz")`` and ``("foo-bar", "baz")`` produce distinct
 names even when their joined prefixes collide.
 
-Orphan cleanup and idempotency rely on identity labels, not container names
-alone. Existing containers keep their pre-change names after upgrade; only new
-deployments receive hashed names.
+Naming logic is shared via ``nemo_platform_plugin.k8s_naming`` (plugins cannot
+import ``nmp_common``). Orphan cleanup and idempotency rely on identity labels,
+not container names alone.
 """
 
 from __future__ import annotations
 
-import hashlib
-import re
-
 from nemo_deployments_plugin.constants import MANAGED_BY_LABEL
+from nemo_platform_plugin.k8s_naming import k8s_safe_name, workspace_name_identity
 
 MANAGED_BY_KEY = "managed-by"
 DEPLOYMENT_WORKSPACE_LABEL = "nemo.nvidia.com/deployment-workspace"
@@ -29,43 +27,10 @@ CONFIG_NAME_LABEL = "nemo.nvidia.com/deployment-config"
 VOLUME_WORKSPACE_LABEL = "nemo.nvidia.com/volume-workspace"
 VOLUME_NAME_LABEL = "nemo.nvidia.com/volume-name"
 
-_HASH_SUFFIX_LENGTH = 8
-
-
-def k8s_safe_name(
-    base_name: str,
-    *,
-    max_length: int = 63,
-    suffix: str = "",
-    hash_input: str | None = None,
-) -> str:
-    """Generate a DNS-label-safe name (RFC 1035) with a mandatory hash suffix.
-
-    Normalizes ``base_name`` for display, truncates when needed, then returns
-    ``{normalized}-{hash8}{suffix}``. The hash is SHA-256 of ``hash_input`` when
-    provided, otherwise ``base_name``. Callers that represent a workspace/name
-    pair should pass ``hash_input=deployment_key(workspace, name)`` so the hash
-    reflects the unambiguous identity rather than the join-ambiguous prefix.
-    """
-    hash_source = hash_input if hash_input is not None else base_name
-    hash_suffix = hashlib.sha256(hash_source.encode()).hexdigest()[:_HASH_SUFFIX_LENGTH]
-    normalized = re.sub(r"[^a-z0-9-]", "-", base_name.lower())
-    normalized = re.sub(r"[-]+", "-", normalized)
-    if normalized and not normalized[0].isalpha():
-        normalized = f"x{normalized}"
-    normalized = normalized.rstrip("-")
-
-    reserved = len(suffix) + len(hash_suffix) + 1
-    if len(normalized) + reserved > max_length:
-        normalized = normalized[: max_length - reserved].rstrip("-")
-    if not normalized:
-        normalized = "x"
-    return f"{normalized}-{hash_suffix}{suffix}"
-
 
 def deployment_key(workspace: str, name: str) -> str:
     """Return the canonical identity string used for hashing and label keys."""
-    return f"{workspace}/{name}"
+    return workspace_name_identity(workspace, name)
 
 
 def container_name(workspace: str, deployment_name: str) -> str:
